+++ Asp.Net Core api with Ef Core SQL & Authorization/ServiceLayer/Services/IEntityService.cs
@@ -8,11 +8,11 @@ namespace $safeprojectname$.Services
 {
     public interface IEntityService
     {
-        Task<T> GetByIdAsync<T>(int entityId) where T : Entity;
+        Task<T> GetByIdAsync<T>(long entityId) where T : Entity;
         Task<T> GetEntity<T>(Expression<Func<T, bool>> filter) where T : Entity;
         Task<IEnumerable<T>> GetAllEntitiesAsync<T>() where T : Entity;
         Task<IEnumerable<T>> GetAllEntitiesAsync<T>(Expression<Func<T, bool>> filter) where T : Entity;
-        Task<T> RemoveEntityAsync<T>(int entityId) where T : Entity;
+        Task<T> RemoveEntityAsync<T>(long entityId) where T : Entity;
         Task<T> CreateAsync<T>(T entity) where T : Entity;
         Task<IEnumerable<T>> CreateRangeAsync<T>(IEnumerable<T> entities) where T : Entity;
         Task<T> UpdateAsync<T>(T entity) where T : Entity;
